==> 06_attempt/05_train_ensemble_model.R <==
# Regression AirBnB Problem ----
# Train & explore ensemble model

# Load package(s) ----
library(tidymodels)
library(tidyverse)
library(here)
library(bonsai)
library(stacks)

# Handle common conflicts
tidymodels_prefer()

# Load candidate model info ----
load(here("06_attempt/results/6a_lin_reg_fit.rda"))
load(here("06_attempt/results/6a_bt_tune.rda"))
load(here("06_attempt/results/6a_rf_tune.rda"))

# create data stack ----
reg_data_stacks <- 
  stacks() |> 
  add_candidates(lin_reg_fit) |> 
  add_candidates(rf_tune) |> 
  add_candidates(bt_tune)

# fit the stack ----
# penalty values for blending (set penalty argument when blending)
blend_penalty <- c(10^(-6:-1), 0.5, 1, 1.5, 2)

# blend predictions (tuning step, set seed)
set.seed(9874) # needed bc tuning process happening in background
reg_stack_blend <- 
  reg_data_stacks |> 
  blend_predictions(penalty = blend_penalty, control = control_stack_grid())

# save blended model stack
save(reg_stack_blend, file = here("06_attempt/results/reg_stack_blend.rda"))

# explore the blended model stack ----
reg_stack_blend |> 
  autoplot()

reg_stack_blend |> 
  autoplot(type = "members")

# fit to training set ----
reg_ensemble <- fit_members(reg_stack_blend)

# save trained ensemble model ----
save(reg_ensemble, file = here("06_attempt/results/reg_ensemble.rda"))
